more translations
This version is now almost complete. All remaining typos are fixed, Labels are updated to fit in the UI where max display width is restricted. The default font is updated from 14pt to 12pt  and uses "Segoe UI" instead of legacy "Microsoft Sans Serif" (like in the English version)

The only non translated parts are the text of numbered items in the user agreement (in English): can I translate them too when the reference text is normally in Simplified Chinese ?