Update fr.xml
Once again this is wrong.
The first column is for the ASCII-only characters to encode in XML by the second column (this is a XML requirement) if we want them in a translation

But the translations MAY and WILL WANT to use other more correct (curly) typographical characters than those undesired vertical ASCII punctuations. The prefered quotes are language-dependant such as ‘curly single quotes’ or “curly double quotes” in English, or only « double angle quotes (guillemets) » in French, and the left curly single quote (’) used everywhere in French as an orthographic apostrophe, or „inverted curly double quotes‟ in German.

Chinese or Japanese may have their own prefered 「fullwidth square quotes」or 〝fullwidth curly quotes〞, but the Latin curly quotes (”) should not be used, as they do not properly align on em-square and are confusable with the repeatition quotes (〃)

In general we should avoid ASCII quotes everywhere in the translated UI, so the XML replacements are not needed at all, EXCEPT in technical texts showing a command-line syntax (where only ASCII is recognized).